Add files via upload
I've specifically worked on the admin pages and got all the buttons working, the sign out button is just a redirect though. I've deleted the Id field in some tables. Added the Calendar page, and an event details page, however I'm thinking about making event details a popup on the calendar. I cleaned up the the calendar page, but the design still isn't perfect. I got an event that was scheduled in the database to appear on the right day in the calendar, as well as some of its attributes(however it is definitely a round-about way of doing it and will not work in the long run). I've edited some of the .js files, particularly the values in some of the arrays, but some of those changes are only temporary for testing. On the calendar page, the event details button links to a new page where I was trying to get the associated Event for that day to display its record, however I couldn't quite figure it out. Going to talk to Jody on Monday if I can't figure it out over the weekend. If you guys have any questions about changes I made you can just ask me on discord. P.S. there are no comments where I made changes but I'll fix that when I get a chance.